requirements.txt
matplotlib==3.3.4
numpy==1.21.6
opencv-python-headless==4.6.0.66
Pillow==9.2.0
pydicom==2.2.2
streamlit==1.13.0
streamlit_cropper==0.2.0
tensorflow==2.10.0
tensorflow-io==0.26.0
tf-nightly==2.12.0.dev20221023
